rename: Changing package name board -> post, table name Post -> post
