+++ backend/src/main/java/com/backend/domain/post/entity/RecruitmentStatus.java
@@ -1,4 +1,4 @@
-package com.backend.domain.board.entity;
+package com.backend.domain.post.entity;
 
 public enum RecruitmentStatus {
     OPEN, // 모집 전
